Divide Top.tsx into breakpoints
