+++ src/view/templates/MobileComponents/mob.module.css
@@ -1,0 +1,39 @@
+/* max-width: 768px */
+.topUnderMob {
+  margin: 70px auto;
+  font-size: 20px;
+  text-align: center;
+  width: 200px;
+}
+
+.topdivdivMob {
+  font-weight: 600;
+  will-change: opacity;
+  width: 95%;
+  margin: 0 auto;
+}
+
+.welcom {
+  font-size: 60px;
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 50px 0;
+}
+
+.see_more {
+  text-decoration: underline;
+  margin: 0 auto;
+  font-size: 30px;
+  text-align: center;
+  cursor: pointer;
+  transform: scale(1.0, 1.0);
+  transition: 0.5s all;
+}
+
+.see_more:hover {
+  transform: scale(1.2, 1.2);
+  transition: 0.5s all;
+  color: purple;
+}
