+++ src/view/templates/PC_Components/pc.module.css
@@ -1,0 +1,78 @@
+.topdivdiv {
+  font-weight: 600;
+  will-change: opacity;
+  width: 70%;
+  margin: 0 auto;
+}
+
+.icon_seemore {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 300px;
+  margin: 0 auto;
+}
+
+.icon {
+  font-size: 60px;
+  display: block;
+  margin: 0 auto;
+}
+
+.right_arrow {
+  animation: right_arrow_animation 1s infinite alternate;
+}
+
+@keyframes right_arrow_animation {
+  0% {
+    transform: translateX(20px);
+  }
+  100% {
+    transform: translateX(0px);
+  }
+}
+
+.left_arrow {
+  animation: left_arrow_animation 1s infinite alternate;
+}
+
+@keyframes left_arrow_animation {
+  0% {
+    transform: translateX(-20px);
+  }
+  100% {
+    transform: translateX(0px);
+  }
+}
+
+.welcom {
+  font-size: 60px;
+  margin: 0 auto;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 50px 0;
+}
+
+.see_more {
+  text-decoration: underline;
+  margin: 0 auto;
+  font-size: 30px;
+  text-align: center;
+  cursor: pointer;
+  transform: scale(1.0, 1.0);
+  transition: 0.5s all;
+}
+
+.see_more:hover {
+  transform: scale(1.2, 1.2);
+  transition: 0.5s all;
+  color: purple;
+}
+
+.topUnder {
+  margin: 70px auto;
+  font-size: 20px;
+  text-align: center;
+  width: 400px;
+}
